requirements.txt
pandas>=1.3.0
numpy>=1.21.0
requests>=2.26.0
web3>=5.23.0
tqdm>=4.62.0
matplotlib>=3.4.0
python-dotenv>=0.19.0